+++ src/org/usfirst/frc/team1675/robot/commands/drive/TankDrive.java
@@ -1,33 +1,24 @@
-package org.usfirst.frc.team1675.robot.commands;
+package org.usfirst.frc.team1675.robot.commands.drive;
 
 import org.usfirst.frc.team1675.robot.Robot;
 
-import edu.wpi.first.wpilibj.buttons.JoystickButton;
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
  *
  */
-public class CheeseDrive extends Command {
+public class TankDrive extends Command {
 
-    public CheeseDrive() {
-    	requires (Robot.driveBase);
-        // Use requires() here to declare subsystem dependencies
-        // eg. requires(chassis);
+    public TankDrive() {
     }
 
     // Called just before this Command runs the first time
     protected void initialize() {
-    	
-    	
     }
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-    	double updown = Robot.oi.getDriverLeftYAxis();
-    	double leftright = Robot.oi.getDriverRightXAxis();
-    	Robot.driveBase.setLeftMotors(updown-leftright);
-    	Robot.driveBase.setRightMotors(updown+leftright);
+    	
     }
 
     // Make this return true when this Command no longer needs to run execute()
